# docs/swagger.yaml
basePath: /
definitions:
...
  models.Nationality:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.PatchNationality:
    properties:
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for enriching names
  title: Name Enricher API
  version: "1.0"
paths:
...
  /nationalities:
    get:
      consumes:
      - application/json
      description: Get a list of nationalities with optional filtering
      parameters:
      - description: Nationality ID
        in: query
        name: id
        type: integer
      - description: Nationality name
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved nationality list
          schema:
            items:
              $ref: '#/definitions/models.Nationality'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List nationalities
      tags:
      - nationalities
    post:
      consumes:
      - application/json
      description: Create a new nationality entry in the database
      parameters:
      - description: Nationality object with name field
        in: body
        name: nationality
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created nationality
          schema:
            $ref: '#/definitions/models.Nationality'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new nationality
      tags:
      - nationalities
  /nationalities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a nationality by its ID
      parameters:
      - description: Nationality ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted nationality
          schema:
            $ref: '#/definitions/models.Nationality'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a nationality
      tags:
      - nationalities
    get:
      consumes:
      - application/json
      description: Get a single nationality by its ID
      parameters:
      - description: Nationality ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved nationality
          schema:
            $ref: '#/definitions/models.Nationality'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Nationality not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a nationality by ID
      tags:
      - nationalities
    put:
      consumes:
      - application/json
      description: Update an existing nationality by ID
      parameters:
      - description: Nationality ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nationality update data
        in: body
        name: nationality
        required: true
        schema:
          $ref: '#/definitions/models.PatchNationality'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated nationality
          schema:
            $ref: '#/definitions/models.Nationality'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a nationality
      tags:
      - nationalities
...
swagger: "2.0"